Use contextify for older versions of node
