+++ lib/ember-app.js
@@ -1,7 +1,8 @@
 var fs = require('fs');
 var path = require('path');
 
-var Contextify = require('contextify');
+var createSandbox = require('./sandbox').createSandbox;
+var runInSandbox = require('./sandbox').runInSandbox;
 var SimpleDOM = require('simple-dom');
 var RSVP    = require('rsvp');
 var chalk = require('chalk');
@@ -32,13 +33,10 @@ function EmberApp(options) {
     FastBoot: { require: sandboxRequire }
   });
 
-  var appFile = fs.readFileSync(appFilePath, 'utf8');
-  var vendorFile = fs.readFileSync(vendorFilePath, 'utf8');
-
-  sandbox.run(vendorFile, vendorFilePath);
+  runInSandbox(vendorFilePath, sandbox);
   debug("vendor file evaluated");
 
-  sandbox.run(appFile, appFilePath);
+  runInSandbox(appFilePath, sandbox);
   debug("app file evaluated");
 
   var AppFactory = sandbox.require('~fastboot/app-factory');
@@ -71,42 +69,6 @@ EmberApp.prototype.visit = function(url) {
     }
   });
 };
-
-function createSandbox(dependencies) {
-  var wrappedConsole =  Object.create(console);
-  wrappedConsole.error = function() {
-    console.error.apply(console, Array.prototype.map.call(arguments, function(a) {
-      return typeof a === 'string' ? chalk.red(a) : a;
-    }));
-  };
-
-  var sandbox = {
-    // Expose the console to the FastBoot environment so we can debug
-    console: wrappedConsole,
-
-    // setTimeout and clearTimeout are an assumed part of JavaScript environments. Expose it.
-    setTimeout: setTimeout,
-    clearTimeout: clearTimeout,
-
-    // Convince jQuery not to assume it's in a browser
-    module: { exports: {} },
-
-    URL: require("url")
-  };
-
-  for (var dep in dependencies) {
-    sandbox[dep] = dependencies[dep];
-  }
-
-  // Set the global as `window`.
-  sandbox.window = sandbox;
-  sandbox.window.self = sandbox;
-
-  // The sandbox is now a JavaScript context O_o
-  Contextify(sandbox);
-
-  return sandbox;
-}
 
 function buildWhitelistedRequire(whitelist, distPath) {
   return function(moduleName) {
