+++ lib/server.js
@@ -101,13 +101,13 @@ function readPackageJSON(distPath) {
     var manifest = pkg.fastboot.manifest;
 
     return {
-      appFile:  path.join(distPath, manifest.appFile),
+      appFile: path.join(distPath, manifest.appFile),
       vendorFile: path.join(distPath, manifest.vendorFile),
       htmlFile: path.join(distPath, manifest.htmlFile),
       moduleWhitelist: pkg.fastboot.moduleWhitelist
     };
   } catch (e) {
-    console.log("Couldn't find %s. You may need to update your version of ember-cli-fastboot.", pkgPath);
+    console.error("Couldn't find %s. You may need to update your version of ember-cli-fastboot.", pkgPath);
   }
 }
